requirements.txt
pillow
html5lib
beautifulsoup4
requests